init: initialize the project using vite
