--- tasks/.husky/pre-commit
@@ -1,1 +1,0 @@
-cd tasks
