--- tasks/package.json
@@ -1,48 +1,0 @@
-{
-    "name": "task-1",
-    "private": true,
-    "version": "0.0.0",
-    "type": "module",
-    "scripts": {
-        "dev": "vite",
-        "build": "tsc -b && vite build",
-        "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-        "test": "vitest",
-        "coverage": "vitest run --coverage",
-        "format:fix": "prettier . --write",
-        "preview": "vite preview",
-        "prepare": "cd .. && husky tasks/.husky"
-    },
-    "dependencies": {
-        "@reduxjs/toolkit": "^2.2.6",
-        "axios": "^1.7.2",
-        "react": "^18.3.1",
-        "react-dom": "^18.3.1",
-        "react-redux": "^9.1.2",
-        "react-router-dom": "^6.24.1"
-    },
-    "devDependencies": {
-        "@testing-library/dom": "^10.3.2",
-        "@testing-library/jest-dom": "^6.4.6",
-        "@testing-library/react": "^16.0.0",
-        "@types/jest": "^29.5.12",
-        "@types/react": "^18.3.3",
-        "@types/react-dom": "^18.3.0",
-        "@typescript-eslint/eslint-plugin": "^7.13.1",
-        "@typescript-eslint/parser": "^7.13.1",
-        "@vitejs/plugin-react": "^4.3.1",
-        "@vitest/coverage-v8": "^2.0.3",
-        "eslint": "^8.57.0",
-        "eslint-config-prettier": "^9.1.0",
-        "eslint-plugin-prettier": "^5.1.3",
-        "eslint-plugin-react-compiler": "^0.0.0-experimental-0998c1e-20240625",
-        "eslint-plugin-react-hooks": "^4.6.2",
-        "eslint-plugin-react-refresh": "^0.4.7",
-        "husky": "^9.1.1",
-        "jsdom": "^24.1.0",
-        "prettier": "^3.3.2",
-        "typescript": "^5.5.3",
-        "vite": "^5.3.1",
-        "vitest": "^2.0.3"
-    }
-}
