--- tasks/src/App.css
@@ -1,185 +1,0 @@
-#root {
-    text-align: center;
-    height: 100vh;
-}
-
-.layout {
-    display: flex;
-    position: relative;
-    flex-direction: column;
-}
-
-.layout.light {
-    background-color: #ffffff;
-}
-
-.layout.dark {
-    background-color: #690375;
-}
-
-.top-section {
-    height: 200px;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    border-radius: 0 0 10% 10%;
-}
-
-.top-section.light {
-    background-color: #a3a7f8;
-}
-
-.top-section.dark {
-    background-color: #2c0e37;
-}
-
-.bottom-section {
-    display: flex;
-}
-
-.bottom-section.light {
-    background-color: #ffffff;
-}
-
-.bottom-section.dark {
-    background-color: #690375;
-}
-
-.left-section {
-    flex: 1;
-    padding: 20px;
-}
-
-.right-section {
-    flex: 1;
-    padding: 20px;
-    border-left: 1px solid #ccc;
-}
-
-.right-section.dark {
-    border-left: 1px solid #2c0e37;
-}
-
-.search-bar {
-    width: 400px;
-    margin-top: 10px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 10px;
-}
-
-.search-input {
-    width: 100%;
-    border-radius: 8px;
-    border: 1px solid transparent;
-    padding: 0.6em 1.2em;
-    font-size: 1em;
-    font-weight: 500;
-    font-family: inherit;
-    background-color: white;
-    color: #646cff;
-}
-
-.search-input:hover {
-    border-color: #646cff;
-}
-
-.search-input:focus,
-.search-input:focus-visible {
-    outline: 4px auto -webkit-focus-ring-color;
-}
-
-h2 {
-    font-size: 1em;
-    line-height: 1.1;
-}
-
-p {
-    font-size: 0.8em;
-}
-
-ul {
-    list-style: none;
-    margin: 0;
-    padding: 0;
-}
-
-.info-message {
-    font-size: 1.5rem;
-}
-
-.error-message {
-    color: red;
-    font-size: 1.5rem;
-}
-
-.search-btn {
-    width: 100px;
-    border-radius: 8px;
-    border: 1px solid transparent;
-    padding: 0.6em 1.2em;
-    font-size: 1em;
-    font-weight: 500;
-    font-family: inherit;
-    cursor: pointer;
-}
-
-.search-btn:hover {
-    border-color: #646cff;
-}
-
-.search-btn:focus,
-.search-btn:focus-visible {
-    outline: 4px auto -webkit-focus-ring-color;
-}
-
-.search-input.dark {
-    background-color: #2c0e37;
-    color: #ffffff;
-    border-color: #ffffff;
-}
-
-.search-input.dark:hover {
-    border-color: #ffffff;
-}
-
-.search-btn.dark {
-    background-color: #2c0e37;
-    color: #ffffff;
-    border-color: #ffffff;
-}
-
-.checkbox {
-    background-color: white;
-}
-
-.product-button {
-    color: #646cff;
-    font-size: 16px;
-    background-color: transparent;
-    border: none;
-}
-
-.selected-items {
-    position: fixed;
-    bottom: 10px;
-    left: 10px;
-    background-color: #7f7f7f;
-    border-radius: 10px;
-}
-
-.selected-items-text {
-    color: white;
-    font-size: 14px;
-}
-
-.selected-items-button {
-    background-color: #646cff;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    padding: 5px 10px;
-    cursor: pointer;
-}
