--- tasks/src/components/ErrorBoundary/ErrorBoundary.tsx
@@ -1,34 +1,0 @@
-import { Component, ErrorInfo, ReactNode } from 'react'
-
-interface ErrorBoundaryState {
-    hasError: boolean
-}
-
-interface ErrorBoundaryProps {
-    children: ReactNode
-}
-
-class ErrorBoundary extends Component<ErrorBoundaryProps, ErrorBoundaryState> {
-    constructor(props: ErrorBoundaryProps) {
-        super(props)
-        this.state = { hasError: false }
-    }
-
-    static getDerivedStateFromError(): ErrorBoundaryState {
-        return { hasError: true }
-    }
-
-    componentDidCatch(error: Error, errorInfo: ErrorInfo) {
-        console.error('ErrorBoundary caught an error', error, errorInfo)
-    }
-
-    render() {
-        if (this.state.hasError) {
-            return <h1>Something went wrong.</h1>
-        }
-
-        return this.props.children
-    }
-}
-
-export default ErrorBoundary
