--- tasks/src/main.tsx
@@ -1,14 +1,0 @@
-import ReactDOM from 'react-dom/client'
-import App from './App.tsx'
-import './index.css'
-import { BrowserRouter } from 'react-router-dom'
-import { store } from './store/store'
-import { Provider } from 'react-redux'
-
-ReactDOM.createRoot(document.getElementById('root')!).render(
-    <BrowserRouter>
-        <Provider store={store}>
-            <App />
-        </Provider>
-    </BrowserRouter>
-)
