--- tasks/src/pages/NotFound/NotFound.tsx
@@ -1,10 +1,0 @@
-function NotFound() {
-    return (
-        <div>
-            <h2>404 - Page Not Found</h2>
-            <p>The page you are looking for does not exist.</p>
-        </div>
-    )
-}
-
-export default NotFound
