--- tasks/tsconfig.json
@@ -1,16 +1,0 @@
-{
-    "files": [],
-    "compilerOptions": {
-        // ... other options
-        "types": ["@testing-library/jest-dom"]
-        // ... other options
-    },
-    "references": [
-        {
-            "path": "./tsconfig.app.json"
-        },
-        {
-            "path": "./tsconfig.node.json"
-        }
-    ]
-}
