--- tasks/tsconfig.node.json
@@ -1,13 +1,0 @@
-{
-    "compilerOptions": {
-        "composite": true,
-        "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.node.tsbuildinfo",
-        "skipLibCheck": true,
-        "module": "ESNext",
-        "moduleResolution": "bundler",
-        "allowSyntheticDefaultImports": true,
-        "strict": true,
-        "noEmit": true
-    },
-    "include": ["vite.config.ts"]
-}
