--- tasks/vite.config.ts
@@ -1,12 +1,0 @@
-import { defineConfig } from 'vitest/config'
-import react from '@vitejs/plugin-react'
-
-// https://vitejs.dev/config/
-export default defineConfig({
-    plugins: [react()],
-    test: {
-        environment: 'jsdom',
-        setupFiles: ['src/_tests/setup.ts'],
-        globals: true,
-    },
-})
